Add useful methods to IsisLevel and test
